redbuttegarden/requirements.txt
anyascii==0.3.0
argcomplete==1.12.3
arrow==0.14.7
asgiref==3.8.1
attrs==21.2.0
Authlib==1.6.6
beautifulsoup4==4.9.3
boto3==1.38.23
botocore==1.38.23
certifi==2024.7.4
cfn-flip==1.3.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.0.3
coverage==7.5.3
defusedxml==0.7.1
Django==5.2.11
django-cas-client==1.5.3
django-cors-headers==3.5.0
django-csp==3.7
django-debug-toolbar==4.2.0
django-environ==0.8.1
django-filter==23.3
django-ical==1.9.2
django-ipware==7.0.1
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-ses==3.5.0
django-storages==1.14.6
django-tables2==2.8.0
django-taggit==5.0.0
django-treebeard==4.5.1
djangorestframework==3.15.2
docutils==0.18
draftjs-exporter==2.1.7
drf-spectacular==0.28.0
durationpy==0.5
et-xmlfile==1.1.0
filetype==1.2.0
geojson==2.5.0
gunicorn==23.0.0
hjson==3.0.2
ics==0.7
idna==3.7
importlib-metadata==4.8.2
iniconfig==1.1.1
jdcal==1.4.1
jmespath==0.10.0
kappa==0.6.0
l18n==2020.6.1
laces==0.1.1
MarkupSafe==2.1.3
openpyxl==3.0.10
packaging==24.2
pep517==0.13.0
pillow==10.3.0
pillow_heif==0.16.0
pip-tools==6.4.0
pipdeptree==2.24.0
placebo==0.8.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pyparsing==2.4.7
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
setuptools==78.1.1
simplejson==3.19.3
six==1.16.0
soupsieve==2.3
sqlparse==0.5.0
tablib==3.1.0
TatSu==5.8.3
telepath==0.3.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
troposphere==3.1.1
Unidecode==1.3.2
urllib3==2.6.3
validators==0.30.0
wagtail==7.0.4
wagtail-accessibility==2.0.0
webencodings==0.5.1
Werkzeug==3.1.5
wheel==0.46.2
Willow==1.11.0
wsgi-request-logger==0.4.6
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
zappa==0.61.4
zappa-django-utils==0.4.1
zipp==3.21.0
